Files into memory
I made it so it stores the compressed files in an array in memory (similar to what we did for simple strings).
I didn't touch decompression but compression almost works. The last character is lost for some reason but the rest is good. I have only tested it using a simple txt file with a few lines and a few different characters.